requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
asttokens 
async-timeout==4.0.2
attrs==23.1.0
backcall 
backports.functools-lru-cache 
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama 
comm 
dataclasses-json==0.5.13
debugpy 
decorator 
distlib==0.3.7
dnspython==2.4.2
et-xmlfile==1.1.0
executing 
faiss-cpu==1.7.4
filelock==3.12.2
filetype==1.2.0
Flask==2.3.2
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-search-results==2.4.2
greenlet==2.0.2
idna==3.4
importlib-metadata 
ipykernel 
ipython 
itsdangerous==2.1.2
jedi 
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client 
jupyter_core 
langchain==0.0.242
langsmith==0.0.14
loguru==0.7.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline 
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio 
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging 
pandas==2.0.3
parso 
pickleshare 
Pillow==9.5.0
pinecone-client==2.2.2
pipenv==2023.7.23
platformdirs==3.9.1
prompt-toolkit 
protobuf==4.24.0
psutil 
pure-eval 
pyarrow==12.0.1
pydantic==1.10.12
pydeck==0.8.0
Pygments 
Pympler==1.0.1
pypdf==3.15.0
pyspark==3.4.1
python-dateutil 
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq 
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
six 
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.19
stack-data 
streamlit==1.25.0
streamlit-chat==0.1.1
tabulate==0.9.0
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado 
tqdm==4.65.0
traitlets 
tweepy==4.14.0
typing-inspect==0.9.0
typing_extensions 
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.9.1
urllib3==2.0.4
validators==0.21.2
virtualenv==20.24.2
virtualenv-clone==0.5.7
watchdog==3.0.0
wcwidth 
Werkzeug==2.3.6
wikipedia==1.4.0
win32-setctime==1.1.0
yarl==1.9.2
zipp 
